app/requirements.txt
click==8.1.7
Flask==3.0.3
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3